add validate script
