+++ src/common/private_keys_paths.ts
@@ -1,0 +1,4 @@
+import Path from "path";
+
+export const oldAppPrivateKeysPath = Path.join(__dirname, "../input/old-app-private-keys.csv")
+export const newAppPrivateKeysPath = Path.join(__dirname, "../input/new-app-private-keys.csv")
